Feat/master

+++ resources/js/pages/administration/priority/components/priority-delete-dialog.tsx
@@ -2,25 +2,31 @@
 
 import { ConfirmDialog } from '@/components/confirm-dialog';
 import { Alert, AlertDescription, AlertTitle } from '@/components/ui/alert';
+import { Input } from '@/components/ui/input';
+import { Label } from '@/components/ui/label';
 import { toast } from '@/hooks/use-toast';
-import { useRoles } from '@/pages/roles/context/roles-context';
-import { Role } from '@/pages/roles/data/schema';
+import { usePriorities } from '@/pages/administration/priority/context/priority-context';
+import { Priority } from '@/pages/administration/priority/data/schema';
 import { usePage } from '@inertiajs/react';
 import { IconAlertTriangle } from '@tabler/icons-react';
+import { useState } from 'react';
 
 interface Props {
-    currentRow?: Role;
     open: boolean;
     onOpenChange: (open: boolean) => void;
+    currentRow: Priority;
 }
 
-export function RolesDeleteDialog({ currentRow, open, onOpenChange }: Props) {
-    const { setShouldReload } = useRoles();
+export function PriorityDeleteDialog({ open, onOpenChange, currentRow }: Props) {
+    const [value, setValue] = useState('');
+    const { setShouldReload } = usePriorities();
     const { csrf_token } = usePage().props as unknown as { csrf_token: string };
+
     const handleDelete = () => {
         if (!currentRow) return;
+        if (value.trim() !== currentRow.name) return;
 
-        fetch('/dashboard/roles/delete/' + currentRow.id, {
+        fetch('/dashboard/priority/delete/' + currentRow.id, {
             method: 'DELETE',
             headers: {
                 'Content-Type': 'application/json',
@@ -30,13 +36,13 @@ export function RolesDeleteDialog({ currentRow, open, onOpenChange }: Props) {
             .then((res) => res.json())
             .then((data) => {
                 setShouldReload(true);
-                toast({ title: 'The following user has been deleted:' });
+                toast({ title: 'The following priority has been deleted:' });
             })
             .catch((err) => {
                 console.error(err);
                 toast({
                     title: 'Error',
-                    description: 'Failed to delete role.',
+                    description: 'Failed to delete priority.',
                     variant: 'destructive',
                 });
             })
@@ -48,19 +54,25 @@ export function RolesDeleteDialog({ currentRow, open, onOpenChange }: Props) {
             open={open}
             onOpenChange={onOpenChange}
             handleConfirm={handleDelete}
+            disabled={value.trim() !== currentRow.name}
             title={
                 <span className="text-destructive">
-                    <IconAlertTriangle className="mr-1 inline-block stroke-destructive" size={18} /> Delete Role
+                    <IconAlertTriangle className="mr-1 inline-block stroke-destructive" size={18} /> Delete User
                 </span>
             }
             desc={
                 <div className="space-y-4">
                     <p className="mb-2">
-                        Are you sure you want to delete <span className="font-bold">{currentRow?.name}</span>?
+                        Are you sure you want to delete <span className="font-bold">{currentRow.name}</span>?
                         <br />
-                        This action will permanently remove the user with the role of{' '}
-                        <span className="font-bold">{currentRow?.name?.toUpperCase()}</span> from the system. This cannot be undone.
+                        This action will permanently remove the priority.
                     </p>
+
+                    <Label className="my-2">
+                        Email:
+                        <Input value={value} onChange={(e) => setValue(e.target.value)} placeholder="Enter email to confirm deletion." />
+                    </Label>
+
                     <Alert variant="destructive">
                         <AlertTitle>Warning!</AlertTitle>
                         <AlertDescription>Please be carefull, this operation can not be rolled back.</AlertDescription>
